Move `units` docutils to a separate module
`units.utils` contains utilities for the internals of the units
machinery, but it also contained utilities for dynamically generating
documentation for unit definition modules. This caused an import loop
that has so far been suppressed by only importing unit classes inside
the documentation utility functions. Moving the documentation utilities
to a separate module broke the import loop.
