+++ astropy/cosmology/units.py
@@ -3,7 +3,7 @@
 
 from __future__ import annotations
 
-from astropy.units.utils import generate_unit_summary as _generate_unit_summary
+from astropy.units.docgen import generate_unit_summary as _generate_unit_summary
 
 __all__ = [
     # redshift equivalencies
