+++ astropy/units/photometric.py
@@ -19,7 +19,7 @@ from astropy.constants.si import L_bol0
 
 from . import astrophys, cgs, si
 from .core import UnitBase, def_unit
-from .utils import generate_unit_summary
+from .docgen import generate_unit_summary
 
 __all__ = []  #  Units are added at the end
